src/server/server_app/requirements.txt
flask==2.2.2
flask_cors==3.0.10
python-dotenv==0.21.0
pymysql